Const Variablen hinzugefügt
